fix watching build
