+++ package.json
@@ -2,18 +2,15 @@
   "name": "rally-study-template",
   "version": "0.3.1",
   "scripts": {
-    "build": "rollup -c rollup.config.intermediate.js && rollup -c rollup.config.final.js",
-    "dev": "rollup --config-enable-developer-mode -c rollup.config.intermediate.js && rollup --config-enable-developer-mode -c rollup.config.final.js",
-    "dev-inter": "rollup -w --config-enable-developer-mode -c rollup.config.intermediate.js",
-    "dev-final": "rollup -w --config-enable-developer-mode -c rollup.config.final.js",
+    "build": "rollup -c rollup.config.js",
+    "dev": "rollup -w --config-enable-developer-mode -c rollup.config.js",
+    "watch": "npm-run-all --parallel dev watch-bundled",
     "docs": "jsdoc . -c jsdoc-conf.json",
     "lint": "eslint . && web-ext lint --ignore-files */*/Readability.js */*/*/Readability.js",
     "package": "npm run build && web-ext build --overwrite-dest --ignore-files \"./tests/**/*\" --ignore-files \"./src/**/*\" --ignore-files \"./rollup.config.js\"",
     "start": "npm run build && web-ext run",
     "test-integration": "npm run package && mv web-ext-artifacts/*.zip web-ext-artifacts/study.xpi && mocha --timeout 30000 \"./tests/integration/*.js\"",
-    "watch": "npm-run-all --parallel watch-raw watch-bundled",
-    "watch-raw": "npm-run-all --parallel dev-inter dev-final",
-    "watch-bundled": "web-ext run --watch-file dist/background.js"
+    "watch-bundled": "web-ext run --watch-file dist/background.js dist/aggregateStatistics.worker.js dist/polClassifier.worker.js dist/covidClassifier.worker.js"
   },
   "devDependencies": {
     "@mozilla/rally": "^0.3.0",
